Renames fileio.py to fileio_iOS.py
