+++ decotools/__init__.py
@@ -1,5 +1,5 @@
 
-from .fileio import get_iOS_files
+from .fileio_iOS import get_iOS_files
 from .fileio_android import get_android_files
 from .blob_extraction import extract_blobs, is_hotspot
